exam_backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
pydantic==1.10.13
pymongo==4.6.3
python-dotenv==1.0.1
requests==2.32.3

